Refactor param names for theme switch
